requirements.txt
opencv-python==4.8.0.76
pandas==2.1.4
pytesseract==0.3.10
Pillow==10.1.0
openpyxl==3.1.2
streamlit==1.30.0
flask
